完成拦截器配置及验证

+++ verynavi_user/src/main/java/com/verynavi/user/service/UserService.java
@@ -12,12 +12,14 @@ import org.springframework.data.redis.core.RedisTemplate;
 import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.stereotype.Service;
 import util.IdWorker;
+import util.JwtUtil;
 
 
 import javax.persistence.criteria.CriteriaBuilder;
 import javax.persistence.criteria.CriteriaQuery;
 import javax.persistence.criteria.Predicate;
 import javax.persistence.criteria.Root;
+import javax.servlet.http.HttpServletRequest;
 import java.util.*;
 import java.util.concurrent.TimeUnit;
 
@@ -44,7 +46,11 @@ public class UserService {
     @Autowired
     private BCryptPasswordEncoder encoder;
 
+    @Autowired
+    private HttpServletRequest request;
 
+    @Autowired
+    private JwtUtil jwtUtil;
     /**
      * 查询全部列表
      *
@@ -119,11 +125,33 @@ public class UserService {
     }
 
     /**
-     * 删除
+     * 删除,必须有admin角色才能删除
      *
      * @param id
      */
     public void deleteById(String id) {
+        String token = (String) request.getAttribute("claims_admin");
+        if(token==null || "".equals(token)){
+            throw new RuntimeException("权限不足!");
+        }
+//        String header = request.getHeader("Authorization");
+//        if(header==null || "".equals(header)){
+//            throw new RuntimeException("权限不足！");
+//        }
+//        if(!header.startsWith("Bearer ")){
+//            throw new RuntimeException("权限不足！");
+//        }
+//        //得到token
+//        String token = header.substring(7);
+//        try{
+//            Claims claims = jwtUtil.parseJWT(token);
+//            String roles = (String) claims.get("roles");
+//            if (roles == null || !roles.equals("admin")){
+//                throw new RuntimeException("权限不足！");
+//            }
+//        }catch (Exception e){
+//            throw new RuntimeException("权限不足！");
+//        }
         userDao.deleteById(id);
     }
 
@@ -206,7 +234,12 @@ public class UserService {
     }
 
 
-
+    /**
+     * 登录
+     * @param mobile
+     * @param password
+     * @return
+     */
     public User login(String mobile, String password) {
         User user = userDao.findByMobile(mobile);
         if(user!=null && encoder.matches(password,user.getPassword())){
